Revise device API to more painlessly handle more efficient out-of-order device access
Based on the OpenCL driver's out-of-order design, running the common miner_thread code while the device is busy, instead of when it's idle.
This replaces the device API's scanhash function with 4 new ones: job_prepare, job_start, job_get_results, and job_process_results
Still compatible with old scanhash API to allow for a gradual transition.